WIP/Demo: move stable parts of `crypto` to a crate
Some types in the `crypto` module need to be shared by
mostly-independent crates:

* `primitives - used when constructing `Witness` or `script_sig`, also
  when accessing script/witness elements.
* `addresses` - there the keys are used in an internal representation
  and in constructors

(Serialized) signatures should be also moved but this change is already
huge without that and there's no need to do it in a single commit.
