+++ bitcoin/src/psbt/serialize.rs
@@ -129,6 +129,7 @@ impl Deserialize for ScriptBuf {
 
 impl Serialize for PublicKey {
     fn serialize(&self) -> Vec<u8> {
+        use crate::crypto::key::PublicKeyExt;
         let mut buf = Vec::new();
         self.write_into(&mut buf).expect("vecs don't error");
         buf
